Allow users to configure the db schema
This commit adds a new option called `schema` to the database
configuration.
This option can be used to configure the PostgreSQL schema.

When set, the option `currentSchema` will be set for the database jdbc
datasource.

From the PostgreSQL documentation jdbc documentation:

```
currentSchema = String

Specify the schema (or several schema separated by commas) to be set
in the search-path. This schema will be used to resolve unqualified
object names used in statements over this connection.
```